requirements.txt
pandas==1.5.3
numpy==1.24.3
unidecode==1.3.0
shapely==2.0.1
geopandas==0.12.2
python-dateutil==2.8.2
selenium>=4.21.0
webdriver-manager>=4.0.1
google-cloud-storage==2.10.0
requests==2.31.0
lxml==4.9.2
beautifulsoup4==4.12.2
